+++ packages/orq-rc/RELEASES.md
@@ -8519,3 +8519,13 @@ Based on:
 - [typescript v4.8.0-rc.45] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.45 - packages/orq-rc
+
+## 2026-04-22 09:56:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.46] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.46 - packages/orq-rc